pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "resonance-risk-screening"
version = "0.1.0"
description = "Physics-guided reduced-order screening of harmonic resonance risk in 33/11 kV substations."
readme = "README.md"
requires-python = ">=3.10,<3.13"
license = "MIT"
authors = [
  { name = "Research Team" }
]
dependencies = [
  "numpy>=1.24",
  "pandas>=2.0",
  "scipy>=1.10",
  "scikit-learn>=1.3",
  "statsmodels>=0.14",
  "matplotlib>=3.7",
  "seaborn>=0.13",
  "pyyaml>=6.0",
  "rich>=13.0"
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Science/Research",
  "Topic :: Scientific/Engineering :: Information Analysis",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "License :: OSI Approved :: MIT License"
]

[project.urls]
Homepage = "https://example.org/resonance-risk-screening"
Repository = "https://example.org/resonance-risk-screening/repo"
Documentation = "https://example.org/resonance-risk-screening/docs"

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "ruff>=0.5",
  "mypy>=1.10",
  "pre-commit>=3.7"
]
docs = [
  "sphinx>=7.0",
  "myst-parser>=3.0",
  "furo>=2024.8.6"
]

[project.scripts]
resonance-screen = "resonance_risk_screening.cli:main"
resonance-pipeline = "resonance_risk_screening.pipeline:main"

[tool.hatch.build.targets.wheel]
packages = ["src/resonance_risk_screening"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
markers = [
  "unit: unit tests",
  "integration: integration tests",
  "e2e: end-to-end workflow tests"
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
disallow_untyped_defs = false
no_implicit_optional = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["yaml"]
ignore_missing_imports = true

[tool.coverage.run]
branch = true
source = ["resonance_risk_screening"]
